clarified interface
